add option to directly load & convert eest fixtures
